added users photos and moved functions
