--- getUserPhoto.php
@@ -1,15 +1,0 @@
-<?php
-function getUserPhotoCell($userId)
-{
-    $imagePath = "images/{$userId}.png";
-
-    if (file_exists($imagePath)) 
-    {
-        return "<td><img id='table-image' src='$imagePath' alt='User Photo'></td>";
-    } 
-    else 
-    {
-        return "<td>No Photo</td>";
-    }
-}
-?>
