+++ stylesheets/mainstyle.css
@@ -93,12 +93,12 @@ body
 
 table 
 {
-    width: 80%;
+    width: 100%;
+    height: 100%;
     border-collapse: collapse;
     max-width: 1500px;
+    min-height: 800px;
     position: relative;
-    margin-top: 27vw;
-
 }
 
 table, th, td 
@@ -176,12 +176,20 @@ h1
     text-shadow: none;
 }
 
-@media screen and (max-width: 900px) 
+.PayrollContainer
+{
+    width: 90%;
+    display: flex;
+    justify-content: center;
+    margin-top: 20vh;
+}
+
+@media screen and (max-width: 1000px) 
 {
     table
     {
        padding: 5000px;
-       margin-top: 100vw;
+       margin-top: 150px;
     }
 }
 
@@ -189,6 +197,7 @@ h1
 /* End of employee-data-general styles */
 
 /* ============================================================================================= */
+
 
 /* ============================================================================================= */
